Add  WebSocketMessageHandlerMessageDispatcherTests Integration
